Move to selector pattern for stakeholders & waves
