+++ client/src/app/pages/migration-waves/migration-waves.tsx
@@ -79,8 +79,6 @@ export const MigrationWaves: React.FC = () => {
 
   const { trackers: trackers } = useFetchTrackers();
   const { data: applications } = useFetchApplications();
-  const { tickets } = useFetchTickets();
-  const { stakeholders } = useFetchStakeholders();
 
   const [migrationWaveModalState, setWaveModalState] = React.useState<
     "create" | MigrationWave | null
@@ -220,77 +218,6 @@ export const MigrationWaves: React.FC = () => {
     },
     expansionDerivedState: { isCellExpanded },
   } = tableControls;
-
-  const getApplicationsOwners = (id: number) => {
-    const applicationOwnerIds = applications
-      .filter((application) => application.migrationWave?.id === id)
-      .map((application) => application.owner?.id);
-
-    return stakeholders.filter((stakeholder) =>
-      applicationOwnerIds.includes(stakeholder.id)
-    );
-  };
-
-  const getApplicationsContributors = (id: number) => {
-    const applicationContributorsIds = applications
-      .filter((application) => application.migrationWave?.id === id)
-      .map((application) =>
-        application.contributors?.map((contributor) => contributor.id)
-      )
-      .flat();
-
-    return stakeholders.filter((stakeholder) =>
-      applicationContributorsIds.includes(stakeholder.id)
-    );
-  };
-
-  const getStakeholdersByMigrationWave = (migrationWave: MigrationWave) => {
-    const holderIds = migrationWave.stakeholders.map(
-      (stakeholder) => stakeholder.id
-    );
-    return stakeholders.filter((stakeholder) =>
-      holderIds.includes(stakeholder.id)
-    );
-  };
-
-  const getStakeholderFromGroupsByMigrationWave = (
-    migrationWave: MigrationWave
-  ) => {
-    const groupIds = migrationWave.stakeholderGroups.map(
-      (stakeholderGroup) => stakeholderGroup.id
-    );
-
-    return stakeholders.filter((stakeholder) =>
-      stakeholder.stakeholderGroups?.find((stakeholderGroup) =>
-        groupIds.includes(stakeholderGroup.id)
-      )
-    );
-  };
-
-  const getAllStakeholders = (migrationWave: MigrationWave) => {
-    let allStakeholders: Stakeholder[] = getApplicationsOwners(
-      migrationWave.id
-    );
-
-    getApplicationsContributors(migrationWave.id).forEach((stakeholder) => {
-      if (!allStakeholders.includes(stakeholder))
-        allStakeholders.push(stakeholder);
-    });
-
-    getStakeholdersByMigrationWave(migrationWave).forEach((stakeholder) => {
-      if (!allStakeholders.includes(stakeholder))
-        allStakeholders.push(stakeholder);
-    });
-
-    getStakeholderFromGroupsByMigrationWave(migrationWave).forEach(
-      (stakeholder) => {
-        if (!allStakeholders.includes(stakeholder))
-          allStakeholders.push(stakeholder);
-      }
-    );
-
-    return allStakeholders;
-  };
 
   return (
     <>
@@ -452,9 +379,7 @@ export const MigrationWaves: React.FC = () => {
                               columnKey: "stakeholders",
                             })}
                           >
-                            {getAllStakeholders(
-                              migrationWave
-                            ).length.toString()}
+                            {migrationWave.allStakeholders.length}
                           </Td>
                           <Td
                             width={20}
@@ -551,7 +476,7 @@ export const MigrationWaves: React.FC = () => {
                           >
                             <ExpandableRowContent>
                               {isCellExpanded(migrationWave, "applications") &&
-                              migrationWave.applications.length > 0 ? (
+                              !!migrationWave.applications.length ? (
                                 <WaveApplicationsTable
                                   migrationWave={migrationWave}
                                   removeApplication={removeApplication}
@@ -559,13 +484,9 @@ export const MigrationWaves: React.FC = () => {
                               ) : isCellExpanded(
                                   migrationWave,
                                   "stakeholders"
-                                ) &&
-                                getAllStakeholders(migrationWave).length > 0 ? (
+                                ) && !!migrationWave.allStakeholders.length ? (
                                 <WaveStakeholdersTable
                                   migrationWave={migrationWave}
-                                  stakeholders={getAllStakeholders(
-                                    migrationWave
-                                  )}
                                 />
                               ) : (
                                 isCellExpanded(migrationWave, "status") && (
